### 2025-11-17 #9440 @editorialbot
[PRE REVIEW]: WoodTapper: a Python package for explaining decision tree ensembles
**Submitting author:** @Abdoulaye-SAKHO (<a href="http://orcid.org/0009-0002-0248-4881">Abdoulaye SAKHO</a>)
**Repository:** https://github.com/artefactory/woodtapper
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.12
