+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={5--32},
   year={2001},
   publisher={Springer}
+  doi = {10.1007/978-3-030-56485-8_3}
 }
 
 @article{pedregosa2011scikit,
@@ -23,6 +24,7 @@
   title={Sirus: Stable and interpretable rule set for classification},
   author={B{\'e}nard, Cl{\'e}ment and Biau, G{\'e}rard and Da Veiga, S{\'e}bastien and Scornet, Erwan},
   year={2021}
+  doi={10.1214/20-ejs1792}
 }
 
 
@@ -43,6 +45,7 @@
   number={90},
   pages={5786},
   year={2023}
+  doi={10.21105/joss.05786}
 }
 
 @article{nussberger2022public,
@@ -54,6 +57,7 @@
   pages={5821},
   year={2022},
   publisher={Nature Publishing Group UK London}
+  doi={10.1038/s41467-022-33417-3}
 }
 
 @article{sokol2024interpretable,
@@ -65,6 +69,7 @@
   pages={3102--3140},
   year={2024},
   publisher={Springer}
+  doi={10.1007/s10618-024-01010-5}
 }
 
 @article{Khalilia:2011,
@@ -75,6 +80,7 @@
   pages={1--13},
   year={2011},
   publisher={Springer}
+  doi={10.1186/1472-6947-11-51}
 }
 
 @inproceedings{ex-fraud,
@@ -84,6 +90,7 @@
   pages={117--127},
   year={2016},
   organization={Springer}
+  doi={10.1007/978-3-319-27400-3_11}
 }
 
 @article{ex-churn,
@@ -93,6 +100,7 @@
   volume={12},
   number={1},
   year={2021}
+  doi={10.12720/jait.12.1.29-35}
 }
 
 @inproceedings{sakho2025harnessing,
@@ -102,6 +110,7 @@
   pages={247--264},
   year={2025},
   organization={Springer}
+  doi={10.1007/978-3-032-06118-8_15}
 }
 
 @software{reback2020pandas,
@@ -145,6 +154,7 @@
   title={Generalized random forests},
   author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},
   year={2019}
+  doi={10.1214/18-aos1709}
 }
 
 @article{skgrf,
