### 2025-11-17 ec0156b
FIX dois typos
