+++ paper/paper.bib
@@ -6,8 +6,8 @@
   number={1},
   pages={5--32},
   year={2001},
-  publisher={Springer}
-  doi = {10.1007/978-3-030-56485-8_3}
+  publisher={Springer},
+  doi={10.1007/978-3-030-56485-8_3}
 }
 
 @article{pedregosa2011scikit,
@@ -23,7 +23,7 @@
 @article{benard2021sirus-classif,
   title={Sirus: Stable and interpretable rule set for classification},
   author={B{\'e}nard, Cl{\'e}ment and Biau, G{\'e}rard and Da Veiga, S{\'e}bastien and Scornet, Erwan},
-  year={2021}
+  year={2021},
   doi={10.1214/20-ejs1792}
 }
 
@@ -44,7 +44,7 @@
   volume={8},
   number={90},
   pages={5786},
-  year={2023}
+  year={2023},
   doi={10.21105/joss.05786}
 }
 
@@ -56,7 +56,7 @@
   number={1},
   pages={5821},
   year={2022},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
   doi={10.1038/s41467-022-33417-3}
 }
 
@@ -68,7 +68,7 @@
   number={5},
   pages={3102--3140},
   year={2024},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/s10618-024-01010-5}
 }
 
@@ -79,7 +79,7 @@
   volume={11},
   pages={1--13},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1186/1472-6947-11-51}
 }
 
@@ -89,7 +89,7 @@
   booktitle={Advances in Nature and Biologically Inspired Computing: Proceedings of the 7th World Congress on Nature and Biologically Inspired Computing (NaBIC2015) in Pietermaritzburg, South Africa, held December 01-03, 2015},
   pages={117--127},
   year={2016},
-  organization={Springer}
+  organization={Springer},
   doi={10.1007/978-3-319-27400-3_11}
 }
 
@@ -99,7 +99,7 @@
   journal={Journal of advances in information technology},
   volume={12},
   number={1},
-  year={2021}
+  year={2021},
   doi={10.12720/jait.12.1.29-35}
 }
 
@@ -109,7 +109,7 @@
   booktitle={Joint European Conference on Machine Learning and Knowledge Discovery in Databases},
   pages={247--264},
   year={2025},
-  organization={Springer}
+  organization={Springer},
   doi={10.1007/978-3-032-06118-8_15}
 }
 
@@ -153,7 +153,7 @@
 @article{grf,
   title={Generalized random forests},
   author={Athey, Susan and Tibshirani, Julie and Wager, Stefan},
-  year={2019}
+  year={2019},
   doi={10.1214/18-aos1709}
 }
 
