### 2025-11-17 b2a6443
FACTORIZE paper to reach words limit
