### 2025-11-17 9a5a48c
Update paper/paper.md
Co-authored-by: gemini-code-assist[bot] <176961590+gemini-code-assist[bot]@users.noreply.github.com>
