+++ paper/paper.md
@@ -129,7 +129,7 @@ Finally, the $l$ training samples with the highest $w(x,x_i)$ values, along with
 The $\textit{skgrf}$ [@skgrf] package is an interface for using the R implementation of generalized random forest in Python. $\textit{skgrf}$ has a specific number of classifiers for specifics learning tasks (causal inference, quantile regression,...). For each task, the user can compute the kernel weights, which are equivalent to our leaf frequency match introduce above. Thus, we aim at comparing the kernel weights derivation from $\textit{skgrf}$ to our $\texttt{ExampleExplanation}$ module. We stress on the fact that our $\texttt{ExampleExplanation}$ is designed for usual tree-based models such as random forest of extra trees and not specifically in a context of causal inference or quantile regression. Thus, the tree building (splitting criterion) of our forest are different from the ones from $\textit{skgrf}$.
 
 ## Implementation and running time
-Our Python implementation of $\texttt{ExampleExplanation}$ adheres to the scikit-learn interface and is agnostic to the underlying tree ensemble (\ref{tab:comparison-grf}). The standard $\texttt{fit}$ and $\texttt{predict}$ methods remain unchanged, while an additional $\texttt{explaination}$ method provides example-based explanations for new samples. We also have implemented a method to load an already trained tree-based model into an $\texttt{ExampleExplanation}$ classifier.
+Our Python implementation of $\texttt{ExampleExplanation}$ adheres to the scikit-learn interface and is agnostic to the underlying tree ensemble (\ref{tab:comparison-grf}). The standard $\texttt{fit}$ and $\texttt{predict}$ methods remain unchanged, while an additional $\texttt{explanation}$ method provides example-based explanations for new samples. We also have implemented a method to load an already trained tree-based model into an $\texttt{ExampleExplanation}$ classifier.
 
 : **Comparison of GRF weight computations in several Python packages.**\label{tab:comparison-grf}
 
