### 2025-11-17 d6a20b9
Factorize paper
