### 2025-11-18 b6b622c
FIX latex rendering
