### 2025-11-18 cdaf364
REFAC: compressed a bit
