### 2025-11-21 d7aa0f5
FIX: print for catgeorical features with more than 2 modalities
