### 2025-11-21 725703a
FIX make binary boolean on features more readble and robuts
