+++ pyproject.toml
@@ -8,7 +8,7 @@ authors = [
     {name = "Abdoulaye SAKHO", email = "abdoulaye7020@gmail.com"},
     {name = "artefactory", email = "abdoulaye.sakho@artefact.com"},
 ]
-version = "0.0.12"
+version = "0.0.13"
 description = "A Python toolbox for interpretable and explainable tree ensembles."
 readme = "README.md"
 license = "MIT"
