+++ docs/installation.md
@@ -19,6 +19,8 @@ And install the required packages into your environment (conda, mamba or pip):
 pip install -e .[dev,docs]
 ```
 
+**Warning:** If you are a Windows user, you need to have a C/C++ compilator before installing woodtapper.
+
 ## Dependencies
 
 WoodTapper requires the following:
