### 2025-12-08 f098656
FIX emphasize math formulation link
