+++ docs/installation.md
@@ -19,7 +19,7 @@ And install the required packages into your environment (conda, mamba or pip):
 pip install -e .[dev,docs]
 ```
 
-**Warning:** If you are a Windows user, you need to have a C/C++ compilator before installing woodtapper.
+**Warning:** If you are a Windows user, you need to have a C/C++ compiler before installing woodtapper.
 
 ## Dependencies
 
@@ -27,7 +27,6 @@ WoodTapper requires the following:
 
 * Python (>=3.11,<3.13)
 * Numpy (>=2.3.1)
-* Pandas (>=2.3.0)
 * Scikit-learn (>=1.6.1)
 
 ## Contributing
