### 2025-12-08 95b0e42
FIX logo and dark.light logo
