### 2026-01-20 15a8f80
Fix warnings on feature_names_in attribute
