### 2026-01-20 bcd48e7
Add new sections required by JOSS
