### 2026-01-20 ae2b38d
Update research impact statement section
