### 2026-01-21 d6fbb50
Update research impact section
