### 2026-01-21 920d0b7
Add Emmanuel feedbacks on python adaptation
