### 2026-01-21 #9440 @Abdoulaye-SAKHO
Hello @crvernon,

Thank you for sharing the updated JOSS requirements. We’ve revised our submission accordingly and would like to relaunch the pre-review process. However, the issue bot indicates that a “State of the field” section is required, while we don’t see this section in the updated JOSS guidelines or the sample paper. Could you clarify whether this section is still expected and, if so, point us to the current guidance or examples?

Please let us know if you need any additional information.
