### 2026-01-23 #9440 @crvernon
Thank you @Abdoulaye-SAKHO - we are having a discussion on the "state of the field" section notice among our editors now.  I'll let you know as soon as possible when we have reached a decision concerning how this section should appear in the paper.  So no action for you on this for now.
