### 2026-01-28 7ce7fcb
Add doi for XGB and Catboost
