+++ paper/paper.bib
@@ -148,6 +148,16 @@
   booktitle={Proceedings of the 22nd acm sigkdd international conference on knowledge discovery and data mining},
   pages={785--794},
   year={2016}
+  doi={10.48550/arXiv.1603.02754}
+}
+
+@article{prokhorenkova2018catboost,
+  title={CatBoost: unbiased boosting with categorical features},
+  author={Prokhorenkova, Liudmila and Gusev, Gleb and Vorobev, Aleksandr and Dorogush, Anna Veronika and Gulin, Andrey},
+  journal={Advances in neural information processing systems},
+  volume={31},
+  year={2018},
+  doi={10.48550/arXiv.1706.09516}
 }
 
 @article{grf,
