+++ paper/paper.md
@@ -31,9 +31,12 @@ date: 14 November 2025
 bibliography: paper.bib
 ---
 
-# Introduction
+# State of the Field
 
-Interpretable machine learning has become an increasingly critical concern [@nussberger2022public;@sokol2024interpretable] as predictive models are deployed in high-stakes settings such as healthcare [@Khalilia:2011], marketing [@ex-churn] or finance [@ex-fraud;@sakho2025harnessing] which is moreover a regulated sector. While complex models, such as tree-based ensemble methods, often yield strong predictive performance, their opacity can pose challenges for accountability, trust and compliance. Among interpretable models, rule-based methods are especially attractive because they are explained in the form of “if-then” statements, which are often easier to audit and communicate than latent feature transformations.
+Interpretable machine learning has become an increasingly critical concern [@nussberger2022public;@sokol2024interpretable] as predictive models are deployed in high-stakes settings such as healthcare [@Khalilia:2011], marketing [@ex-churn] or finance [@ex-fraud;@sakho2025harnessing] which is moreover a regulated sector.
+Widely used libraries such as scikit-learn [@pedregosa2011scikit], XGBoost [@chen2016xgboost] and CatBoost [@prokhorenkova2018catboost] provide efficient implementations of tree-based models, but their primary focus is predictive performance and training scalability rather than interpretability. In parallel, explanation toolkits such as SHAP and LIME offer model-agnostic or surrogate-based attributions that are broadly applicable, but they can be computationally expensive for large ensembles.
+
+`WoodTapper` complements these ecosystems by treating tree ensembles as first-class objects for interpretability: it provides a dedicated Python toolbox to inspect, decompose, and explain predictions using methods that directly leverage the discrete structure of trees, enabling faithful explanations that are aligned with the model’s internal logic. This focus supports reproducible analysis workflows for researchers who need transparent decision rules, feature interactions grounded in splits, and diagnostic views of ensemble behavior. Rather than extending training libraries (which are optimized for fitting), `WoodTapper` targets the distinct scholarly need of interpretable ensemble analysis, offering functionality that is not sufficiently addressed by general-purpose learners or purely model-agnostic explainers.
 
 
 # Statement of need
@@ -41,13 +44,6 @@ Interpretable machine learning has become an increasingly critical concern [@nus
 The original SIRUS algorithm [@benard2021sirus-classif;@benard2021interpretable-regression] offered a principled approach to generate simple and stable rule-based models from random forests. However, its implementations have been limited to R and Julia [@benard2021sirus-classif;@huijzer2023sirus-jl], creating accessibility barriers for the Python data science community. `WoodTapper` addresses this gap by offering a native Python implementation that integrates with the scikit-learn ecosystem. Furthermore, `WoodTapper` extends rules extraction $(i)$ from all the tree-based models in scikit learn (Random Forest, Gradient Boosting and Extremely Randomized Trees) and $(ii)$ to the multiclass classification setting.
 
 In addition, `WoodTapper` introduces an example-based explainability methodology that can be applied to all scikit-learn tree-based models. This approach associates predicted samples with representative samples from the training data set, explaining tree-based models predictions through examples.
-
-# Software design
-`WoodTapper` package adheres to the scikit-learn [@pedregosa2011scikit] estimator interface.
-This design enables smooth integration with existing workflows involving pipelines, cross-validation, and model selection, and enables to efficiently benefit from future maintenance updates and improvements to scikit-learn.
-The implementation leverages NumPy for numerical computation and joblib for parallel processing to optimize performance on large datasets (\ref{tab:comparison}).
[…]
