+++ paper/paper.bib
@@ -147,7 +147,7 @@
   author={Chen, Tianqi and Guestrin, Carlos},
   booktitle={Proceedings of the 22nd acm sigkdd international conference on knowledge discovery and data mining},
   pages={785--794},
-  year={2016}
+  year={2016},
   doi={10.48550/arXiv.1603.02754}
 }
 
