### 2026-01-28 4d2b9f9
Update state of field with build vs contribute
