### 2026-01-28 dfe12fb
Update Catboost as inproceedings
