+++ paper/paper.bib
@@ -151,11 +151,11 @@
   doi={10.48550/arXiv.1603.02754}
 }
 
-@article{prokhorenkova2018catboost,
+@inproceedings{prokhorenkova2018catboost,
   title={CatBoost: unbiased boosting with categorical features},
   author={Prokhorenkova, Liudmila and Gusev, Gleb and Vorobev, Aleksandr and Dorogush, Anna Veronika and Gulin, Andrey},
-  journal={Advances in neural information processing systems},
-  volume={31},
+  booktitle={Advances in neural information processing systems},
+  pages={6638--6648},
   year={2018},
   doi={10.48550/arXiv.1706.09516}
 }
