### 2026-01-28 #9440 @Abdoulaye-SAKHO
Hello @crvernon  and @arfon , thank you for your response.
We have updated our manuscript in accordance with the new guidelines and would be happy to restart the pre-review process.
Please let us know if you need any additional information.
