### 2026-02-26 #10112 @fangzhou-xie
Hi @Manishms18, thanks for your interest in reviewing JOSS articles. We already have two reviewers for this submission, and we currently are just waiting for both reviewers to finish their reviews.

I would be happy to invite you to review for future submissions in this area, and you could [register yourself as a reviewer (click the "Volunteer to review" on the right)](https://joss.theoj.org/about) to be potentially invited by other editors too. Thanks!
