### 2026-02-26 #10112 @jncraton
@Abdoulaye-SAKHO I've just begun reviewing this, but it looks good so far.

I've attempted to reproduce the diagrams from the paper, but I'm running into issues. Specifically, I just tried to run https://github.com/artefactory/woodtapper/blob/main/paper/exp_py.ipynb, but I'm hitting an issue with this line:

```python
os.mkdir('reproduce-exp', exist_ok=True) # create output directory
```

`exist_ok` doesn't appear to be defined for [os.mkdir](https://docs.python.org/3/library/os.html#os.mkdir). Am I doing something wrong here? Could you check this out on your end and make sure these notebooks are working as expected?
