### 2026-02-27 #10112 @Abdoulaye-SAKHO
Hello @jncraton ,

Thanks again for reviewing our submission.

The issue came from a change I made from `pathlib.Path.mkdir` to `os.mkdir` to better align with the rest of the codebase, and I didn’t verify it properly afterward. I’ve now updated the code to use `os.makedirs`.

Could you please try reproducing the experiments again? Let me know if you need anything else.
