### 2026-03-01 #10112 @jncraton
@Abdoulaye-SAKHO Thank you for addressing that. That part of the notebook is working correctly for me now.

Would it be possible to automate dependency installation, including pinning the versions that you used in the paper in these notebooks? That would help with long-term reproducibility.

My hope would be that I would be able to simply [run these notebooks via colab](https://colab.research.google.com/github/artefactory/woodtapper/blob/main/paper/exp_py.ipynb), but that doesn't work for right now because:

1. `woodtapper` has to be manually installed first. That's relatively simple, but it would be nice to have this dependency automated and tied to the precise version from the paper.
2. Even with `sklearn` installed, I see `AttributeError: module 'sklearn.tree._utils' has no attribute 'WeightedFenwickTree'` when attempting to run this notebook. I haven't look into this deeply, but it appears that the internals of sklearn may have changed at some point and this needs to be run against a specific version.

The same basic issue is present in the R notebook. Instructions state that sirus must be installed, but it would be handy if this were automatic and pinned to the version that was used in the paper to prevent any confusion in the future.
