### 2026-03-03 #10112 @Abdoulaye-SAKHO
Hello @jncraton ,

Thank you for your feedback. We have updated the repository to address the requested changes. In `paper/README.md`, we now include a link to a Google Colab environment for running the experiments. The notebooks have been updated to be compatible with Colab, and we added clear instructions so users can easily run and test the experiments there.
Thank you as well for suggesting this improvement, which significantly strengthens the reproducibility of our experiments. During this update, we also noticed that installing WoodTapper may require recompilation when `scikit-learn` is already installed. We have therefore updated the main `README.md` with guidelines for this situation.

Please let us know if anything is not working as expected with the Colab link, or if you have any additional suggestions.
