### 2026-03-06 #10112 @jncraton
@Abdoulaye-SAKHO Is the code to replicate the performance graphs in the paper available? It obviously wouldn't generate identical results due to system differences, but it would still be helpful to replicate the general trend and could be helpful down the line if researchers want to re-run these comparisons against future versions of these tools.
