### 2026-03-10 b6cf85d
docs:[experiments] add section to infrom that folder timing contains infos for computation time reproductibility
