### 2026-03-10 25a9f03
feat[reproductibility]: add timining reproductibility setup for JOSS paper
