### 2026-03-11 #10112 @Abdoulaye-SAKHO
@jncraton — I created a new branch, `dev_abdoulaye`, that contains all the material required to reproduce the computation-time experiments. If everything looks good to you, I will merge it into `main`.

## 1) ExampleBasedExplainability timings
All computation-time results for *ExampleBasedExplainability* can be reproduced using the following notebooks:
- `paper/timing/timing-skgrf.ipynb`
- `paper/timing/timing-exampleSampling-with-skgrf-plot.ipynb`

**Note:** these `skgrf`-based notebooks cannot be executed on Google Colab, because `skgrf` cannot be installed there.

## 2) ExtractionRules timings
You can now reproduce the experiments from the paper that compare the runtime of the three SIRUS implementations: **R**, **Julia**, and **Python (ours)**.

Since each implementation has its own dependencies, you will need to set up a separate environment for each one; the setup procedure is described in `paper/timing/README.md`.

The provided scripts report **only the total runtime** (forest training + rule extraction). The runtimes of the two intermediate steps—(i) growing the forest and (ii) extracting the rules—are **not measured dynamically** by the scripts; instead, they are **hard-coded** in the `.csv` files under `time-csv/`.

This is because, for the R implementation, we patched the source code (see `R/Sirus.R`, starting at line 172 in the upstream [SIRUS repository](https://gitlab.com/drti/sirus/-/tree/master?ref_type=heads)) to print both the forest-growing time and the rule-extraction time, installed this patched version in our R environment, saved the console output to a `.txt` log, and then parsed these logs to populate the `.csv` files. We followed the same workflow for the Julia implementation.

To facilitate benchmarking on your own setup, we provide a configuration that recomputes **only the total fitting time** and updates the corresponding `.csv` files accordingly. In addition, `paper/timing/README.md` includes guidance on how to patch the R/Julia sources to re-enable the step-wise breakdown (forest growth vs. rule extraction). If you do so on your side, you can regenerate those values and update the relevant `.csv` files to plot your own forest-growing and rule-extraction times.

The step-wise breakdown is included in the paper mainly for pedagogical clarity; the primary metric of interest remains the **total runtime**.

Would these additions address your concern? We tried to strike the best trade-off between reproducibility and implementation constraints.
