### 2026-03-13 #10112 @jncraton
@Abdoulaye-SAKHO As I'm going through the checklist for the paper, I noticed that it doesn't have a separate "Summary" section as required. It might be nice for readers if something like the second paragraph in the current "Statement of need" section were pulled up as a "Summary" section right at the beginning.
