### 2026-03-13 7656578
fix(paper joss): add summary section at the begining of the paper after review suggestion
