### 2026-03-13 c9e334c
fix(paper joss): typos from summary section
