### 2026-03-13 #10112 @Abdoulaye-SAKHO
@jncraton - Indeed, there is no “Summary” section because we thought it was no longer required under the updated guidelines (and it was not flagged by the compilation bot). However, as you suggested, we have moved the second paragraph of the current “Statement of need” section to the beginning of the paper and presented it as a dedicated “Summary” section.
