### 2026-03-13 #10112 @jncraton
@Abdoulaye-SAKHO

> There is no “Summary” section because we thought it was no longer required under the updated guidelines (and it was not flagged by the compilation bot)

I noticed that as well. It sounds like you may have a better understanding of the current guidelines than I do. I only mentioned the summary section because it was listed as required in the checklist. If this is no longer required and you prefer your previous paper structure, I'm quite happy with that.

Overall, I think this project and paper look good. Thank you for your time building, maintaining, and updating this as well as your time writing it up and formalizing your results.
