### 2026-04-16 79c8b71
fix(parameters): set parameters to the ones from the papers
