### 2026-04-16 d917359
fix(parameters): set subfeatures to the same values as those from R and Jl (modified for rapid check)
