### 2026-04-16 77133f7
doc(path): add path to timing folder and readme
