### 2026-04-29 artefactory/woodtapper#37 @jungtaekkim
JOSS Review - Manuscript
I appreciate the authors' efforts to develop this interesting software.

I will provide my comments on a submitted manuscript.

1. In an author list, last names are capitalized.  I believe it is intentional, but to me it looks strange.  If the authors think it is okay to change them, I think the authors can use the standard author name format.

2. In Summary, "We describe these two modules below" needs to be specified.  You can mention where they are described.  Or, you may remove this sentence, if you think it is redundant.

3. Some references should be introduced in the first appearance or be added.  For example, the authors cited Scikit-learn and random forest, but they are cited later instead of the first mention.  Moreover, SHAP, LIME, gradient boosting, extremely randomized trees, and NumPy should be cited.

4. In Line 37, "explaining tree-based models predictions" seems grammatically incorrect.

5. In Line 40, "Table" should be "Tables."

6. In Research impact statement, "200 publications" and "1500 times" should include the dates when the authors refer to.

7. Footnote symbols should be located after a period or a comma.

8. Can you add a brief plain-language description on Rules Extraction Module?  It can be one or two sentences.

9. In Tables 1 and 3, what are "Py" and "Jl"?  I think they are Python and Julia.  I feel like the authors can describe them with full names.  If there is a better way to express them instead of describing them without proper explanation, it would be great.

10. In Line 91, 250GB RAM is not natural to me.  I am not certain but it might be 256GB?

11. In Line 99, the authors argued that both results are very similar.  But, my question is why they are not identical.  I am not saying about an implementation issue.  If there are any factors to make them not identical, please describe them in a few sentences.

12. In Acknowledgements, if you are okay, please follow the standard author name format.

This is part of the review of JOSS submission (https://github.com/openjournals/joss-reviews/issues/10112).
