### 2026-04-29 artefactory/woodtapper#38 @jungtaekkim
JOSS Review - Documentation
Thank you again for your efforts to write documentation part.

I think documentation generally looks great.  However, I have one comment.

For the following links:

https://artefactory.github.io/woodtapper/1_tutorials/
https://artefactory.github.io/woodtapper/2_tutorials_example_exp/

it would be better to use texts instead of images for expected outputs.  Please update them if possible.

This is part of the review of JOSS submission (https://github.com/openjournals/joss-reviews/issues/10112).
