### 2026-04-30 305ab76
feat(citations): add shap, lime, gbdt, etra trees
