### 2026-04-30 artefactory/woodtapper#37 @Abdoulaye-SAKHO
Hello, thank you very much for your detailed review. We have addressed and incorporated all the remarks we were able to resolve.

However, we are not sure we understand comment 4 regarding the grammar of the sentence on line 37 (“explaining tree-based models predictions”). Could you please clarify what change you would recommend?

Additionally, regarding remark 8, we are unsure where to add the description of the Rules Extraction module. Is the current wording in the subsection "Formulation", and if so, where would you prefer us to expand the description?

Thank you again for your time.
