+++ docs/2_tutorials_example_exp.md
@@ -40,12 +40,27 @@ Xy_explain = RFExplained.explanation(X_test) # Get the 5 most similar samples fo
 ```python
 X_test.iloc[0,:]  # First test sample
 ```
-![](images/x-test-0.png)
+!!! example "Output"
+```text
+a    5.8
+b    2.8
+c    5.1
+d    2.4
+Name: 114, dtype: float64
+```
 
 ```python
 Xy_explain[0] # Explanation for the first test sample
 ```
-![](images/Xy-explain-0.png)
+!!! example "Output"
+```text
+(array([[5.8, 2.7, 5.1, 1.9],
+        [5.8, 2.7, 5.1, 1.9],
+        [5.7, 2.5, 5. , 2. ],
+        [5.9, 3. , 5.1, 1.8],
+        [6.3, 3.4, 5.6, 2.4]]),
+ array([2, 2, 2, 2, 2]))
+```
 
 ## Load an existing RandomForestClassifier into the explainer
 
