+++ paper/paper.bib
@@ -183,7 +183,7 @@
 }
 
 @inproceedings{lime,
-  title={" Why should i trust you?" Explaining the predictions of any classifier},
+  title={"Why Should I Trust You?": Explaining the Predictions of Any Classifier},
   author={Ribeiro, Marco Tulio and Singh, Sameer and Guestrin, Carlos},
   booktitle={Proceedings of the 22nd ACM SIGKDD international conference on knowledge discovery and data mining},
   pages={1135--1144},
