### 2026-04-30 b06b026
fix(typos): typos on researc impact section
