### 2026-04-30 #10112 @Abdoulaye-SAKHO
Hello @jungtaekkim , thank you for taking the time to review our submission.

We have addressed both of the issues you raised. The documentation output issue has been fully resolved. Regarding the paper, we believe we have implemented all requested changes, and we added two clarification questions in the issue thread. Please let us know if you would like us to make further updates to the paper.
