### 2026-05-01 artefactory/woodtapper#37 @jungtaekkim
Thank you for your updates!

Regarding comment 4, I believe "tree-based models' predictions" or "tree-based model's predictions" should be correct.  If I am wrong, ignore it.

Regarding comment 8, I meant some brief sentences in the subsection "Formulation" to explain definitions and equations.
