### 2026-05-02 0f90fbd
fix(review): fix typos and describe RulesExtraction Module in Formulation subsection
