### 2026-05-04 0dfb00d
fix(refs): add missings dois for gbdt and shap
