+++ paper/paper.bib
@@ -17,6 +17,7 @@
   volume={12},
   pages={2825--2830},
   year={2011},
+  doi={https://doi.org/10.48550/arXiv.1201.0490},
   publisher={JMLR. org}
 }
 
@@ -179,7 +180,8 @@
   author={Lundberg, Scott M and Lee, Su-In},
   journal={Advances in neural information processing systems},
   volume={30},
-  year={2017}
+  year={2017},
+  doi={10.48550/arXiv.1705.07874}
 }
 
 @inproceedings{lime,
@@ -187,7 +189,8 @@
   author={Ribeiro, Marco Tulio and Singh, Sameer and Guestrin, Carlos},
   booktitle={Proceedings of the 22nd ACM SIGKDD international conference on knowledge discovery and data mining},
   pages={1135--1144},
-  year={2016}
+  year={2016},
+  doi={10.18653/v1/n16-3020}
 }
 
 @article{geurts2006extremely,
@@ -207,7 +210,8 @@
   journal={Annals of statistics},
   pages={1189--1232},
   year={2001},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.1214/aos/1013203451}
 }
 
 @Article{numpy,
