### 2026-05-04 ccc515e
fix(refs):doi from sklearn
