+++ paper/paper.bib
@@ -17,7 +17,7 @@
   volume={12},
   pages={2825--2830},
   year={2011},
-  doi={https://doi.org/10.48550/arXiv.1201.0490},
+  doi={10.48550/arXiv.1201.0490},
   publisher={JMLR. org}
 }
 
