### 2026-05-05 bb8d51d
fix(typos): capitalized P for Python
