+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{pedregosa2011scikit,
-  title={Scikit-learn: Machine learning in Python},
+  title={Scikit-learn: Machine learning in {P}ython},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={the Journal of machine Learning research},
   volume={12},
@@ -169,7 +169,7 @@
 }
 
 @article{skgrf,
-  title={skgrf - grf in Python},
+  title={skgrf - grf in {P}ython},
   author={Christopher Flynn},
   year={2021},
   url={https://github.com/crflynn/skgrf/tree/main?tab=readme-ov-file}
