### 2026-05-05 5eeba1f
Update paper/paper.md
Co-authored-by: gemini-code-assist[bot] <176961590+gemini-code-assist[bot]@users.noreply.github.com>
