### 2026-05-05 3306069
Joss vernon review (#43)
* fix(refs):doi from sklearn

* fix(typos): capitalized P for Python

* fix(typos): brackets for capitilized P

* Update paper/paper.md
