### 2026-05-05 85cbc58
feat(accepted paper): add citation.cff
