### 2026-05-05 2bb84d9
fix(doc): citation paper
