### 2026-05-05 2f6b721
fix(doc): date article citation
Co-authored-by: gemini-code-assist[bot] <176961590+gemini-code-assist[bot]@users.noreply.github.com>
