### 2026-06-15 814f71b
feat(type): add options for returning example samples as df with their targets
