### 2026-06-15 9bce4b9
fix: example based explainability with df option
