+++ docs/2_tutorials_example_exp.md
@@ -34,7 +34,7 @@ RFExplained.fit(X_train,y_train)
 
 ## Generate example-based explainability
 ```python
-Xy_explain = RFExplained.explanation(X_test) # Get the 5 most similar samples for each test sample
+Xy_explain = RFExplained.explanation(X_test,to_pandas=True) # Get the 5 most similar samples for each test sample
 ```
 
 ```python
@@ -53,14 +53,14 @@ Name: 114, dtype: float64
 Xy_explain[0] # Explanation for the first test sample
 ```
 !!! example "Output"
-```text
-(array([[5.8, 2.7, 5.1, 1.9],
-        [5.8, 2.7, 5.1, 1.9],
-        [5.7, 2.5, 5. , 2. ],
-        [5.9, 3. , 5.1, 1.8],
-        [6.3, 3.4, 5.6, 2.4]]),
- array([2, 2, 2, 2, 2]))
-```
+
+| Index | feature_0 | feature_1 | feature_2 | feature_3 | target |
+| :--- | :---: | :---: | :---: | :---: | :---: |
+| 0 | 5.8 | 2.7 | 5.1 | 1.9 | 2 |
+| 1 | 5.8 | 2.7 | 5.1 | 1.9 | 2 |
+| 2 | 5.7 | 2.5 | 5.0 | 2.0 | 2 |
+| 3 | 5.9 | 3.0 | 5.1 | 1.8 | 2 |
+| 4 | 6.3 | 3.4 | 5.6 | 2.4 | 2 |
 
 ## Load an existing RandomForestClassifier into the explainer
 
