### 2026-06-15 9f042e6
feat(infos): add paper link and citation
