### 2026-06-15 6cabe13
feat(infos): add paper link and citation
