### 2026-06-15 29895bd
fix(dependencies): align sk-learn version to the compiled one with cython
