+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=78.1.1", "wheel", "Cython>=3.1.2","numpy>=2.3.1","scikit-learn>=1.6.1" ]
+requires = ["setuptools>=78.1.1", "wheel", "Cython>=3.1.2","numpy>=2.3.1","scikit-learn==1.6.1" ]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -22,7 +22,7 @@ classifiers = [
 requires-python = ">=3.11,<3.13"
 dependencies = [
     "numpy>=2.3.1",
-    "scikit-learn>=1.6.1",
+    "scikit-learn==1.6.1",
 ]
 
 [project.urls]
