### 2026-06-15 e16eddf
fix: faster explanation df option
