### 2026-06-15 db0d462
feat(exp): add multiclass compatibility
