### 2026-06-15 5f8ac78
Feat faster explanation (#47)
* fix(refs):doi from sklearn

* fix: faster explanation df option

* feat(exp): add multiclass compatibility

* feat: add df explanation
