### 2026-06-15 8ea6d3c
feat: add features names options for example samples
