+++ woodtapper/example_sampling/base.py
@@ -135,7 +135,9 @@ class ExplanationMixin:
             leafs_by_sample, self.train_samples_leaves, leaf_sizes
         )
 
-    def explanation(self, X, n_examples=5, batch_size=None, to_pandas=False):
+    def explanation(
+        self, X, n_examples=5, batch_size=None, to_pandas=False, feature_names=None
+    ):
         """
         Explanation procedure.
         Show the 5 most similar samples based on the frequency of training samples ending in the same leaf as the new sample
@@ -164,7 +166,8 @@ class ExplanationMixin:
         # Get the 5 most similar samples
         if to_pandas:
             list_ = []
-            feature_names = [f"feature_{j}" for j in range(self.train_X.shape[1])]
+            if feature_names is None:
+                feature_names = [f"feature_{j}" for j in range(self.train_X.shape[1])]
             for i in range(most_similar_idx.shape[0]):
                 df = pd.DataFrame(
                     self.train_X[most_similar_idx[i]], columns=feature_names
