+++ woodtapper/example_sampling/base.py
@@ -168,6 +168,11 @@ class ExplanationMixin:
             list_ = []
             if feature_names is None:
                 feature_names = [f"feature_{j}" for j in range(self.train_X.shape[1])]
+            elif len(feature_names) != self.train_X.shape[1]:
+                raise ValueError(
+                    f"Length of feature_names ({len(feature_names)}) must match "
+                    f"the number of features ({self.train_X.shape[1]})."
+                )
             for i in range(most_similar_idx.shape[0]):
                 df = pd.DataFrame(
                     self.train_X[most_similar_idx[i]], columns=feature_names
@@ -179,6 +184,7 @@ class ExplanationMixin:
                 else:
                     for j in range(self.train_y.shape[1]):
                         df[f"target_{j}"] = self.train_y[most_similar_idx[i]][:, j]
+                df.reset_index(drop=True, inplace=True)
                 list_.append(df)
             return list_
         else:
