### 2026-06-15 3c8f9ff
feat: add features names options for example samples (#48)
* feat: add features names options for example samples

* remove index

* fix: features names check
