### 2026-08-17 2298d8d
feat: verbosity attribute for rules extractors (false by default)
