### 2026-08-17 0ca0409
fix(doc): add c compilation errors guidelines
