### 2026-08-17 14cc788
fix(doc): preicse doctrings oin the extension of each class to obtain example based explainability
