+++ woodtapper/example_sampling/classification_explanation.py
@@ -1,7 +1,8 @@
 """
+Example-based explanations for classification tree-based models.
 
-ExampleExplanation for classification.
-
+This module provides wrapper classes that extend scikit-learn classifiers
+with example-based explainability capabilities through the ExplanationMixin.
 """
 
 from sklearn.ensemble import (
@@ -13,12 +14,24 @@ from .base import ExplanationMixin
 
 
 class RandomForestClassifierExplained(ExplanationMixin, RandomForestClassifier):
-    """ExplanationExample RandomForestClassifier"""
+    """Random Forest classifier with example-based explanations.
+
+    Extends RandomForestClassifier with methods for generating interpretable
+    instance-level explanations.
+    """
 
 
 class ExtraTreesClassifierExplained(ExplanationMixin, ExtraTreesClassifier):
-    """ExplanationExample ExtraTreesClassifier"""
+    """Extremely Randomized Trees classifier with example-based explanations.
+
+    Extends ExtraTreesClassifier with methods for generating interpretable
+    instance-level explanations.
+    """
 
 
 class GradientBoostingClassifierExplained(ExplanationMixin, GradientBoostingClassifier):
-    """ExplanationExample GradientBoostingClassifier"""
+    """Gradient Boosting classifier with example-based explanations.
+
+    Extends GradientBoostingClassifier with methods for generating interpretable
+    instance-level explanations.
+    """
