### 2026-08-17 7b9bfee
feat: verbosity attribute for rules extractors (false by default) (#49)
* feat: verbosity attribute for rules extractors (false by default)

* fix(doc): add c compilation errors guidelines

* fix(doc): preicse doctrings oin the extension of each class to obtain example based explainability
