### 2026-08-18 ae4f7bc
fix: add license stament, affilitions
