+++ docs/installation.md
@@ -4,12 +4,12 @@ To install the required packages in a virtual environment, run the following com
 
 # 🛠 Installation
 
-## From PyPi:
+## From PyPI:
 ```bash
 pip install woodtapper
 ```
-> **Warning (scikit-learn already installed):**
-> If you install `woodtapper` in an environment where `scikit-learn` is already present, the prebuilt PyPI wheel may not be compatible with your existing `scikit-learn` binary. In that case, reinstall `woodtapper` from source so it is compiled against the `scikit-learn` version in your environment:
+> **Compatibility note:**
+> WoodTapper includes a compiled Cython component that depends on the binary interface of `scikit-learn`. If `scikit-learn` is already installed in your environment, the prebuilt PyPI wheel may be incompatible with it. WoodTapper is currently pinned to `scikit-learn` 1.6.1: versions 1.9 and newer are not compatible with this component. To use another compatible version, rebuild WoodTapper locally with:
 >
 > ```bash
 > pip uninstall -y woodtapper
@@ -32,9 +32,9 @@ WoodTapper requires the following:
 
 * Python (>=3.11,<3.13)
 * Numpy (>=2.3.1)
-* Scikit-learn (>=1.6.1)
+* Scikit-learn (==1.6.1)
 
-## Contributing
+# Contributing
 You are welcome to contribute to the project ! You can help in various ways:
 
 * raise issues
@@ -45,31 +45,3 @@ You are welcome to contribute to the project ! You can help in various ways:
 * develop new tests
 
 We recommend to first open an [issue](https://github.com/artefactory/woodtapper/issues) to discuss your ideas.
-
-## 📜 Citation
-
-If you find the code useful, please consider citing us:
-
-```bibtex
-@article{Sakho2026,
-doi = {10.21105/joss.10112},
-url = {https://doi.org/10.21105/joss.10112},
-year = {2026}, publisher = {The Open Journal},
-volume = {11},
-number = {121},
-pages = {10112},
-author = {Sakho, Abdoulaye and Aouad, Jad and Gauthier, Carl-Erik and Malherbe, Emmanuel and Scornet, Erwan},
-title = {WoodTapper: a Python package for explaining decision tree ensembles},
-journal = {Journal of Open Source Software} }
-```
-For SIRUS methodology, consider citing:
-```bibtex
-@article{benard2021sirus,
-  title={Sirus: Stable and interpretable rule set for classification},
-  author={Benard, Clement and Biau, Gerard and Da Veiga, Sebastien and Scornet, Erwan},
-  year={2021}
-}
-
-### License
-
-The use of this software is under the MIT license.
