### 2026-08-18 7daddb8
fix(docs):  guidelines for Cython compilation errors and add affiliations/License (#50)
* fix: typo from md

* fix: docstrings for regression example sampling

* fix: compilation error guidelines

* fix: add license stament, affilitions
