+++ woodtapper/example_sampling/regression_explanation.py
@@ -13,12 +13,24 @@ from .base import ExplanationMixin
 
 
 class RandomForestRegressorExplained(ExplanationMixin, RandomForestRegressor):
-    """ExplanationExample RandomForestRegressor"""
+    """ExplanationExample RandomForestRegressor
+
+    Extends RandomForestRegressor with methods for generating interpretable
+    instance-level explanations.
+    """
 
 
 class ExtraTreesRegressorExplained(ExplanationMixin, ExtraTreesRegressor):
-    """ExplanationExample ExtraTreesRegressor"""
+    """ExplanationExample ExtraTreesRegressor
+
+    Extends ExtraTreesRegressor with methods for generating interpretable
+    instance-level explanations.
+    """
 
 
 class GradientBoostingRegressorExplained(ExplanationMixin, GradientBoostingRegressor):
-    """ExplanationExample GradientBoostingRegressor"""
+    """ExplanationExample GradientBoostingRegressor
+
+    Extends GradientBoostingRegressor with methods for generating interpretable
+    instance-level explanations.
+    """
